refactor: modularize simulation into modules
